+++ changelog.md
@@ -1,5 +1,8 @@
 # Mailcode Changelog
 
+## v3.7.2 - Command Escaping
+- Syntax: Added escaping whole commands with `\{command\}` syntax.
+
 ## v3.7.1 - HubL Date Fix
 - Velocity: Removed HubL exception triggered for Velocity-specific date formats.
 - HubL: Updated `showdate` with a conditional date format command.